Update components and views folders structure
