+++ package.json
@@ -13,6 +13,7 @@
     "pinia": "^2.0.23",
     "roboto-fontface": "*",
     "vue": "^3.2.38",
+    "vue-i18n": "^9.2.2",
     "vue-router": "4",
     "vuetify": "^3.0.0",
     "webfontloader": "^1.0.0"
