+++ src/components/Button/index.ts
@@ -1,1 +1,1 @@
-export {default as Button} from './component.vue'
+export {default as Button} from './Component.vue'
