--- src/components/SideNavLogo/index.ts
@@ -1,1 +1,0 @@
-export {default as SideNavLogo} from './component.vue'
