--- src/components/Toolbar/index.ts
@@ -1,1 +1,0 @@
-export {default as Toolbar} from './component.vue'
