--- src/components/ToolbarItem/component.vue
@@ -1,14 +1,0 @@
-<script setup lang="ts">
-</script>
-
-<template>
-  <div class="toolbar-item">
-    <slot />
-  </div>
-</template>
-
-<style scoped lang="scss">
-.toolbar-item {
-  display: flex;
-}
-</style>
