--- src/components/ToolbarItem/index.ts
@@ -1,1 +1,0 @@
-export {default as ToolbarItem} from './component.vue'
