+++ src/components/index.ts
@@ -1,14 +1,12 @@
-export * from './Toolbar'
-export * from './ToolbarItem'
 export * from './Button'
-export * from './SideNav'
-export * from './SideNavItem'
-export * from './SideNavLogo'
 export * from './Avatar'
 export * from './Menu'
 export * from './Card'
 export * from './Icon'
 export * from './TextField'
 export * from './SearchInput'
-export * from './ToolbarMenu'
 export * from './UserExternalLinks'
+export * from './DataTable'
+export * from './List'
+export * from './ListItem'
+export * from './ListItemTitle'
