+++ src/plugins/index.ts
@@ -9,11 +9,13 @@ import { loadFonts } from './webfontloader'
 import vuetify from './vuetify'
 import pinia from '../store'
 import router from '../router'
+import i18n from './translations'
+
 
 // Types
 import type { App } from 'vue'
 
 export function registerPlugins(app: App) {
   loadFonts()
-  app.use(vuetify).use(router).use(pinia)
+  app.use(vuetify).use(router).use(pinia).use(i18n)
 }
