+++ src/store/app.ts
@@ -3,6 +3,6 @@ import { defineStore } from 'pinia'
 
 export const useAppStore = defineStore('app', {
   state: () => ({
-    //
+    userName: '',
   }),
 })
