+++ src/views/Account/index.ts
@@ -1,0 +1,1 @@
+export {default as Account} from './Component.vue'
