+++ src/views/Admin/Component.vue
@@ -1,0 +1,6 @@
+<template>
+  <h1>Admin</h1>
+</template>
+
+<script lang="ts" setup>
+</script>
