--- src/views/Admin/Home.vue
@@ -1,7 +1,0 @@
-<template>
-  <h1>Admin</h1>
-</template>
-
-<script lang="ts" setup>
-
-</script>
