+++ src/views/Admin/index.ts
@@ -1,0 +1,1 @@
+export {default as Admin} from './Component.vue'
