--- src/views/Dashboard/Home.vue
@@ -1,7 +1,0 @@
-<template>
-  <h1>Dashboard</h1>
-</template>
-
-<script lang="ts" setup>
-
-</script>
