+++ src/views/Dashboard/_components/Table/index.ts
@@ -1,0 +1,1 @@
+export {default as Table} from './component.vue'
