+++ src/views/Home/Component.vue
@@ -1,0 +1,7 @@
+<template>
+  <h1>Home</h1>
+</template>
+
+<script lang="ts" setup>
+
+</script>
