--- src/views/Home/Home.vue
@@ -1,7 +1,0 @@
-<template>
-  <h1>HOME</h1>
-</template>
-
-<script lang="ts" setup>
-
-</script>
