+++ src/views/Navigation/SideNav/NavigationDrawer/index.ts
@@ -1,0 +1,1 @@
+export {default as NavigationDrawer} from './Component.vue'
