+++ src/views/Navigation/SideNav/index.ts
@@ -1,0 +1,1 @@
+export {default as SideNav} from './Component.vue'
