+++ src/views/Navigation/Toolbar/_components/ToolbarItem/Component.vue
@@ -1,0 +1,11 @@
+<template>
+  <div class="toolbar-item">
+    <slot />
+  </div>
+</template>
+
+<style scoped lang="scss">
+.toolbar-item {
+  display: flex;
+}
+</style>
