+++ src/views/Navigation/Toolbar/index.ts
@@ -1,0 +1,1 @@
+export {default as Toolbar} from './Component.vue'
