+++ src/views/Navigation/index.ts
@@ -1,0 +1,1 @@
+export {default as Navigation} from './Component.vue'
